moving taskdefinitiontype and reducing some warnings
