+++ src/task_definition/definition_factory_test.rs
@@ -1,0 +1,4 @@
+#[test]
+fn test() {
+    assert_eq!(1, 2);
+}
