+++ src/task_definition/mod.rs
@@ -2,10 +2,14 @@ mod bash_task;
 mod dummy_task;
 mod python_task;
 mod task_def;
+mod definition_factory;
 
 pub use bash_task::BashTaskDefinition;
 pub use dummy_task::DummyTaskDefinition;
 pub use python_task::PythonTaskDefinition;
 pub use task_def::{
-    generate_task_definition_id, string_to_definition_type, TaskDefinition, TaskDefinitionType,
+    generate_task_definition_id, TaskDefinition
+};
+pub use definition_factory::{
+    string_to_definition_type, TaskDefinitionType
 };
