+++ src/task_definition/task_def.rs
@@ -1,27 +1,10 @@
 use crate::errors::YoshiError;
 use crate::task_output::TaskOutput;
 use crate::type_definition::TaskId;
+use crate::task_definition::TaskDefinitionType;
 use dyn_clone::DynClone;
 use std::collections::HashMap;
 use std::fmt::Debug;
-
-/// Enum identifying the variant of Definition
-#[derive(Debug, PartialEq)]
-pub enum TaskDefinitionType {
-    Bash,
-    Python,
-    Dummy,
-}
-
-/// Given a string, return an enum that link to a definition variant
-pub fn string_to_definition_type(def_name: String) -> Option<TaskDefinitionType> {
-    match def_name.as_str() {
-        "python_task_definition" => Some(TaskDefinitionType::Python),
-        "bash_task_definition" => Some(TaskDefinitionType::Bash),
-        "dummy_task_definition" => Some(TaskDefinitionType::Dummy),
-        _ => None,
-    }
-}
 
 /// Trait that define a task that can be started
 /// basically what's to be done
