+++ src/task_node_test.rs
@@ -1,5 +1,5 @@
-use crate::runners::{FakeTaskRunner, TaskRunner, TaskRunnerType};
-use crate::task_definition::{generate_task_definition_id, BashTaskDefinition};
+use crate::runners::TaskRunnerType;
+use crate::task_definition::BashTaskDefinition;
 use crate::task_instance::{TaskInstance, TaskStatus};
 use crate::task_node::TaskNode;
 use crate::task_output::TaskOutput;
